% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\name{distance_clump}
\alias{distance_clump}
\title{Clumping SNPs to identify lead SNPs in association signals}
\usage{
distance_clump(
  dat,
  chr_col,
  bp_col,
  p_col,
  distance_window = 1e+06,
  index_threshold = 5e-08
)
}
\arguments{
\item{dat}{data.table containing summary statistics}

\item{chr_col}{character string containing the column name for chromosome}

\item{bp_col}{character string containing the column name for position}

\item{p_col}{character string containing the column name for p-value}

\item{distance_window}{numeric value specifying the distance window for clumping}

\item{index_threshold}{numeric value specifying the p-value threshold for candidate SNPs}
}
\value{
data.table with clumped SNPs
}
\description{
This function performs clumping of SNPs based on their physical distance and p-values.
}
\author{
Tom Willis
}
