% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{Draw Q-Q plot.}
\usage{
qqplot(daf, p_col)
}
\arguments{
\item{daf}{data.frame containing p-values.}

\item{p_col}{p-value column label}
}
\value{
ggplot object
}
\description{
Draws a Q-Q plot of p-values from a data.frame.
}
